Refactor: Extract base class
