+++ src/test/php/web/auth/unittest/BasicTest.class.php
@@ -2,12 +2,12 @@
 
 use lang\IllegalStateException;
 use unittest\Assert;
-use web\auth\BasicAuthentication;
+use web\auth\Basic;
 use web\filters\Invocation;
 use web\io\{TestInput, TestOutput};
 use web\{Handler, Request, Response};
 
-class BasicAuthenticationTest {
+class BasicTest {
   const REALM = 'Testing';
 
   private $login;
@@ -21,12 +21,12 @@ class BasicAuthenticationTest {
 
   #[Test]
   public function can_create() {
-    new BasicAuthentication(self::REALM, $this->login);
+    new Basic(self::REALM, $this->login);
   }
 
   #[Test]
   public function required() {
-    $auth= new BasicAuthentication(self::REALM, $this->login);
+    $auth= new Basic(self::REALM, $this->login);
     Assert::instance(Handler::class, $auth->required(function($req, $res) {
       // ...
     }));
@@ -34,7 +34,7 @@ class BasicAuthenticationTest {
 
   #[Test]
   public function yields_401_and_www_authenticate_without_authorization() {
-    $auth= new BasicAuthentication(self::REALM, $this->login);
+    $auth= new Basic(self::REALM, $this->login);
 
     $req= new Request(new TestInput('GET', '/'));
     $res= new Response(new TestOutput());
@@ -50,7 +50,7 @@ class BasicAuthenticationTest {
 
   #[Test, Values(['Bearer TOKEN', 'Basic NOT.BASE64', 'Basic Tk9fQ09MT04=', 'Basic dGVzdDpJTkNPUlJFQ1Q', 'Basic SU5DT1JSRUNUOnNlY3JldA=='])]
   public function yields_401_and_www_authenticate_with_incorrect($authorization) {
-    $auth= new BasicAuthentication(self::REALM, $this->login);
+    $auth= new Basic(self::REALM, $this->login);
 
     $req= new Request(new TestInput('GET', '/', ['Authorization' => $authorization]));
     $res= new Response(new TestOutput());
@@ -66,7 +66,7 @@ class BasicAuthenticationTest {
 
   #[Test]
   public function passes_user() {
-    $auth= new BasicAuthentication(self::REALM, $this->login);
+    $auth= new Basic(self::REALM, $this->login);
 
     $req= new Request(new TestInput('GET', '/', ['Authorization' => 'Basic dGVzdDpzZWNyZXQ=']));
     $res= new Response(new TestOutput());
